+++ pyA20/i2c/i2c_lib.h
@@ -26,7 +26,7 @@
 
 extern int i2c_open(char *device, uint8_t address);
 extern int i2c_close(int fd);
-extern int i2c_send(int fd, uint8_t *buffer, uint8_t num_bytes);
-extern int i2c_read(int fd, uint8_t *buffer, uint8_t num_bytes);
+extern int i2c_send(int fd, uint8_t *buffer, uint32_t num_bytes);
+extern int i2c_read(int fd, uint8_t *buffer, uint32_t num_bytes);
 
 #endif
